+++ src/lib/Scenes/Search2/components/SearchListItem.tsx
@@ -1,17 +1,22 @@
-import { themeGet } from "@styled-system/theme-get"
 import OpaqueImageView from "lib/Components/OpaqueImageView/OpaqueImageView"
-import { ArtworkIcon, AuctionIcon, Box, CloseIcon, Flex, Sans, Spacer, Touchable, useSpace } from "palette"
+import { CloseIcon, Flex, Spacer, Touchable, useSpace } from "palette"
 import React from "react"
-import { Pressable, ScrollView } from "react-native"
-import styled from "styled-components/native"
 
-export const SearchListItem: React.FC<any> = ({
-  onPress,
-  imageURL,
+interface SearchListItemProps {
+  categoryName: string
+  imageURL: string | null
+  onPress: () => void
+  onDelete?: () => void
+  InfoComponent: React.ComponentType<any>
+}
+
+export const SearchListItem: React.FC<SearchListItemProps> = ({
   categoryName,
-  Info,
+  imageURL,
+  onPress,
   onDelete,
-  showNavigationButtons,
+  InfoComponent,
+  children,
 }) => {
   const space = useSpace()
 
@@ -25,7 +30,7 @@ export const SearchListItem: React.FC<any> = ({
           />
           <Spacer ml={1} />
           <Flex style={{ flex: 1 }}>
-            <Info />
+            <InfoComponent />
           </Flex>
           {!!onDelete && (
             <Touchable
@@ -45,58 +50,7 @@ export const SearchListItem: React.FC<any> = ({
         </Flex>
       </Touchable>
 
-      {showNavigationButtons && (
-        <>
-          <Spacer m={0.5} />
-
-          <Flex flexDirection="row" alignItems="center">
-            <ScrollView
-              horizontal
-              contentContainerStyle={{ paddingHorizontal: space(1) }}
-              showsHorizontalScrollIndicator={false}
-            >
-              <Spacer ml={1} />
-
-              <Pressable onPress={() => onPress({ artistTab: "Artworks" })}>
-                {({ pressed }) => (
-                  <QuickNavigationButton>
-                    <Box mr={0.5}>
-                      <ArtworkIcon fill={pressed ? "blue100" : "black100"} />
-                    </Box>
-                    <Sans size="3" color={pressed ? "blue100" : "black100"}>
-                      Artworks
-                    </Sans>
-                  </QuickNavigationButton>
-                )}
-              </Pressable>
-
-              <Spacer ml={1} />
-
-              <Pressable onPress={() => onPress({ artistTab: "Insights" })}>
-                {({ pressed }) => (
-                  <QuickNavigationButton>
-                    <Box mr={0.5}>
-                      <AuctionIcon fill={pressed ? "blue100" : "black100"} />
-                    </Box>
-                    <Sans size="3" color={pressed ? "blue100" : "black100"}>
-                      Auction Results
-                    </Sans>
-                  </QuickNavigationButton>
-                )}
-              </Pressable>
-            </ScrollView>
-          </Flex>
-        </>
-      )}
+      {children}
     </>
   )
 }
-
-const QuickNavigationButton = styled(Flex)`
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  padding: 2px 10px;
-  border: 1px solid ${themeGet("colors.black30")};
-  border-radius: 20;
-`
